requirements.txt
Flask==3.1.2
flask_cors==6.0.1
google_api_python_client==2.187.0
google_auth_oauthlib==1.2.3
protobuf==6.33.2
PyJWT==2.10.1
python_dateutil==2.8.2
pywebview==6.1
Requests==2.32.5
selenium==4.39.0
selenium_wire==5.1.0
webdriver_manager==4.0.2
